app/requirements.txt
fastapi==0.119.0
python-jose==3.5.0
python-dotenv==1.1.1
passlib==1.7.4
Jinja2==3.1.6
python-multipart==0.0.20
bcrypt==4.0.1
uvicorn==0.38.0
